pyproject.toml
[tool.poetry]
name = "course-7-homework"
version = "0.1.0"
description = ""
authors = ["Zidolch <yury.batin@gmail.com>"]
readme = "README.md"
packages = [{include = "course_7_homework"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
mypy = "^1.2.0"
pydantic = "^1.10.7"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
